+++ slider/src/test/java/io/monstarlab/mosaic/slider/ValueCheckPointDistributionTest.kt
@@ -1,5 +1,6 @@
 package io.monstarlab.mosaic.slider
 
+import io.monstarlab.mosaic.slider.distribution.ValueCheckPointDistribution
 import org.junit.Assert.*
 
 import org.junit.After
